third_party/blink/web_tests/http/tests/devtools/a11y-axe-core/application-panel/preview-json-image-a11y-test-expected.txt

Tests accessibility in the preview view for JSON and images using the axe-core linter.

Initial state:
Application
  Manifest (selected)
  Service Workers
  Clear storage
Storage
  Local Storage
    http://127.0.0.1:8000
  Session Storage
    http://127.0.0.1:8000
  IndexedDB
  Web SQL
  Cookies
    http://127.0.0.1:8000
Cache
  Cache Storage
  Application Cache
Background Services
  Background Fetch
  Background Sync
Frames
  top
visible view: unknown

Running: testJSONView
Revealed json-value.js:
Application
  Manifest
  Service Workers
  Clear storage
Storage
  Local Storage
    http://127.0.0.1:8000
  Session Storage
    http://127.0.0.1:8000
  IndexedDB
  Web SQL
  Cookies
    http://127.0.0.1:8000
Cache
  Cache Storage
  Application Cache
Background Services
  Background Fetch
  Background Sync
Frames
  top
    Scripts
      json-value.js (selected)
    inspected-page.html
visible view: json
aXe violations: []


Running: testImageView
Revealed image.png:
Application
  Manifest
  Service Workers
  Clear storage
Storage
  Local Storage
    http://127.0.0.1:8000
  Session Storage
    http://127.0.0.1:8000
  IndexedDB
  Web SQL
  Cookies
    http://127.0.0.1:8000
Cache
  Cache Storage
  Application Cache
Background Services
  Background Fetch
  Background Sync
Frames
  top
    Images
      image.png (selected)
    Scripts
      json-value.js
    inspected-page.html
visible view: image
aXe violations: []
